(PA-7366) Install released gems
This makes it possible to install puppet and facter from the rubygesm-puppetcore
server, while still using rubygems.org for everything else.

The setup-ruby action with bundler-cache: true will perform a bundle install, so
we need to make the credentials available to the action.

This also adds a step to list installed gems to verify the expected gems are
installed.